Refactoring project structure, Changed AutoLoggedComponent to interface
