+++ src/main/java/ru/ithub/util/FCCAPIUtil.java
@@ -9,7 +9,6 @@ import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.time.Duration;
 
 /**
  * Free Currency Converter API
