+++ src/main/java/ru/ithub/util/logging/AutoLoggedComponent.java
@@ -2,8 +2,6 @@ package ru.ithub.util.logging;
 
 import java.util.logging.Logger;
 
-public abstract class AutoLoggedComponent {
-    public Logger getLogger() {
-        return Logger.getGlobal();
-    }
+public interface AutoLoggedComponent {
+    Logger getLogger();
 }
