+++ src/main/java/ru/ithub/util/logging/AutoLoggedComponentBuilder.java
@@ -1,5 +1,7 @@
 package ru.ithub.util.logging;
 
+import ru.ithub.annotation.Parameter;
+
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -8,21 +10,19 @@ import java.util.Arrays;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 
-public class AutoLoggedComponentBuilder<INTERFACE, IMPLEMENTATION extends AutoLoggedComponent> {
-    private Class<INTERFACE> interfaceClass;
-    private Class<IMPLEMENTATION> implementationClass;
+public class AutoLoggedComponentBuilder < INTERFACE, IMPLEMENTATION extends AutoLoggedComponent > {
+    private Class < INTERFACE > interfaceClass;
+    private Class < IMPLEMENTATION > implementationClass;
 
+    public AutoLoggedComponentBuilder() {}
 
-    public AutoLoggedComponentBuilder() {
-     }
-
-    public AutoLoggedComponentBuilder setInterfaceClass(Class<INTERFACE> interfaceClass) {
+    public AutoLoggedComponentBuilder setInterfaceClass(Class < INTERFACE > interfaceClass) {
         this.interfaceClass = interfaceClass;
 
         return this;
     }
 
-    public AutoLoggedComponentBuilder setImplementationClass(Class<IMPLEMENTATION> implementationClass) {
+    public AutoLoggedComponentBuilder setImplementationClass(Class < IMPLEMENTATION > implementationClass) {
         this.implementationClass = implementationClass;
 
         return this;
@@ -30,31 +30,43 @@ public class AutoLoggedComponentBuilder<INTERFACE, IMPLEMENTATION extends AutoLo
 
     public Object build() throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
         return Proxy.newProxyInstance(
-            Thread.currentThread().getContextClassLoader(),
-            new Class<?>[]{
-                    interfaceClass
-            },
-            new InvocationHandler() {
-                final IMPLEMENTATION target = implementationClass.getDeclaredConstructor().newInstance();
+                Thread.currentThread().getContextClassLoader(),
+                new Class<?> [] {
+                        interfaceClass
+                },
+                new InvocationHandler() {
+                    final IMPLEMENTATION target = implementationClass.getDeclaredConstructor().newInstance();
 
-                @Override
-                public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-                    try {
-                        Object returnValue = method.invoke(target, args);
+                    @Override
+                    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+                        try {
+                            Object returnValue = method.invoke(target, args);
+                            if (isAutoLoggedComponent(implementationClass)) {
+                                if (returnValue != null) {
+                                    target.getLogger().log(Level.INFO, methodToString(method) + " >> " + returnValue);
+                                } else {
+                                    target.getLogger().log(Level.INFO, methodToString(method));
+                                }
+                            }
 
-                        if (returnValue != null) {
-                            target.getLogger().log(Level.INFO, methodToString(method) + " >> " + returnValue);
-                        } else {
-                            target.getLogger().log(Level.INFO, methodToString(method));
+                            return returnValue;
+
+                        } catch (InvocationTargetException ite) {
+                            throw ite.getCause();
                         }
-
-                        return returnValue;
-                    } catch (InvocationTargetException ite) {
-                        throw ite.getCause();
                     }
                 }
-            }
         );
+    }
+
+    private boolean isAutoLoggedComponent(Class < IMPLEMENTATION > implementationClass) {
+        for (Class<?> implemented : implementationClass.getInterfaces()) {
+            if (implemented.getName().equals(AutoLoggedComponent.class.getName())) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     private String methodToString(Method method) {
@@ -67,8 +79,8 @@ public class AutoLoggedComponentBuilder<INTERFACE, IMPLEMENTATION extends AutoLo
                                 return "";
                             }
                         })
-                        .collect(Collectors.toSet())
-                + ")";
+                        .collect(Collectors.toSet()) +
+                ")";
 
         result = result.replace("[", "").replace("]", "");
         return result;
